restructure main
