+++ src/utils/modules.tsx
@@ -1,7 +1,3 @@
-import Welcome from '../components/Welcome';
-import NotImplementedYet from '../components/NotImplementedYet';
-import Explore from '../components/Explore/Explore';
-
 export type moduleIds = (
     "welcome" |
     "browse" |
@@ -11,11 +7,3 @@ export type moduleIds = (
 )
 
 export type changeModuleFunction = (newModule: moduleIds) => void;
-
-export const moduleComponents = {
-    welcome: <Welcome />,
-    browse: <Explore />,
-    request: <NotImplementedYet />,
-    access: <NotImplementedYet />,
-    login: <NotImplementedYet />,
-}
